pbl_app_backend/requirements.txt
-i https://pypi.org/simple
absl-py==2.1.0; python_version >= '3.7'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
astunparse==1.6.3
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
cli-helpers[styles]==2.3.1; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
configobj==5.0.9; python_version >= '3.7'
cryptography==44.0.0; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0; python_version >= '3.8'
fastapi[standard]==0.115.5; python_version >= '3.8'
fastapi-cli[standard]==0.0.5; python_version >= '3.8'
filelock==3.16.1; python_version >= '3.8'
flatbuffers==24.3.25
gast==0.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gdown==5.2.0; python_version >= '3.8'
google-pasta==0.2.0
greenlet==3.1.1; python_version >= '3.7'
grpcio==1.68.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
h5py==3.12.1; python_version >= '3.9'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
jwt==1.3.1; python_version >= '3.6'
keras==3.7.0; python_version >= '3.9'
libclang==18.1.1
markdown==3.7; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
ml-dtypes==0.4.1; python_version >= '3.9'
namex==0.0.8
numpy==2.0.2; python_version >= '3.9'
opencv-python==4.10.0.84; python_version >= '3.6'
opt-einsum==3.4.0; python_version >= '3.8'
optree==0.13.1; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pgcli==4.1.0; python_version >= '3.8'
pgspecial==2.1.3
pillow==11.0.0; python_version >= '3.9'
prompt-toolkit==3.0.48; python_full_version >= '3.7.0'
protobuf==5.29.0; python_version >= '3.8'
psycopg[binary]==3.2.3; python_version >= '3.8'
psycopg-binary==3.2.3; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.1; python_version >= '3.8'
pydantic-core==2.27.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.19; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
requests[socks]==2.32.3; python_version >= '3.8'
retina-face==0.0.17; python_full_version >= '3.5.5'
rich==13.9.4; python_full_version >= '3.8.0'
setproctitle==1.3.4; python_version >= '3.8'
setuptools==75.6.0; python_version >= '3.9'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
sqlalchemy==2.0.36; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
starlette==0.41.3; python_version >= '3.8'
tabulate[widechars]==0.9.0; python_version >= '3.7'
tensorboard==2.18.0; python_version >= '3.9'
tensorboard-data-server==0.7.2; python_version >= '3.7'
tensorflow==2.18.0; python_version >= '3.9'
termcolor==2.5.0; python_version >= '3.9'
tf-keras==2.18.0; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typer==0.13.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
uvicorn[standard]==0.32.1; python_version >= '3.8'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==0.24.0; python_version >= '3.8'
wcwidth==0.2.13
websockets==14.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
wheel==0.45.1; python_version >= '3.8'
wrapt==1.17.0; python_version >= '3.8'
